Allow reusing bam/sam files with conditions

+++ modules/Bio/EnsEMBL/EGPipeline/SequenceAlignment/ShortRead/AlignSequence.pm
@@ -74,10 +74,24 @@ sub run {
   my $seq_file_2  = $self->param('seq_file_2');
   my $clean_up    = $self->param_required('clean_up');
   
-  my $sam_file = "$seq_file_1.sam";
-  $aligner->align($genome_file, $sam_file, $seq_file_1, $seq_file_2);
-  my $bam_file = $aligner->sam_to_bam($sam_file);
-  unlink $sam_file if $clean_up;
+  my $sam_file   = "$seq_file_1.sam";
+  my $bam_file   = "$seq_file_1.sam";
+  my $bam_exists = -s $bam_file;
+  my $sam_exists = -s $sam_file;
+  
+  # Can we reuse some files?
+  if (not $bam_exists) {
+    unlink $sam_file if -s $sam_file;
+    $aligner->align($genome_file, $sam_file, $seq_file_1, $seq_file_2);
+  } elsif ($sam_exists) {
+    unlink $bam_file;
+  }
+  
+  # In all cases except for when we already had a bam file: convert the sam to bam
+  if (not $bam_exists or $sam_exists) {
+    my $bam_file = $aligner->sam_to_bam($sam_file);
+    unlink $sam_file if $clean_up;
+  }
   
   my $index_cmds = $self->param('index_cmds') || [];
   my $align_cmds = $aligner->align_cmds;
